update lume/variable and solid-js with support for better ESM standards defined by Node
This is a breaking change because the new support is in solid-js 0.24
and that has some API modifications. This means that components made
with this version of lume/element may not be interoperable with those
made with previous lume/element because compile output will differ be
using slightly different APIs, even if as a user the end usage is
identical.